// src/app/model/products-reserve.ts
export interface ProductsReserve {
  carsInsuranceCode?: string;
  carsInsuranceDesc?: string;
  carsProductsId?: string;
  productsResActive?: string;
  productsResAsOfDate?: Date;
  productsResBodily?: number;
  productsResId?: string;
  productsResInsured?: number;
  productsResMaterial?: number;
  productsResTp?: number;
  sysCreatedBy?: string;
  sysCreatedDate?: Date;
  sysUpdatedBy?: string;
  sysUpdatedDate?: Date;
}